+++ src/utils/data_buffer.cpp
@@ -4,7 +4,7 @@ DataBuffer::DataBuffer (int num_samples, size_t buffer_size)
 {
     this->buffer_size = buffer_size;
     this->num_samples = num_samples;
-    data = new float[buffer_size * num_samples];
+    data = new double[buffer_size * num_samples];
     timestamps = new double[buffer_size];
     first_free = first_used = count = 0;
 }
@@ -20,11 +20,11 @@ bool DataBuffer::is_ready ()
     return (data != NULL);
 }
 
-void DataBuffer::add_data (double timestamp, float *value)
+void DataBuffer::add_data (double timestamp, double *value)
 {
     lock.lock ();
     this->timestamps[first_free] = timestamp;
-    memcpy (this->data + first_free * num_samples, value, sizeof (float) * num_samples);
+    memcpy (this->data + first_free * num_samples, value, sizeof (double) * num_samples);
     first_free = next (first_free);
     count++;
     if (first_free == first_used)
@@ -35,27 +35,27 @@ void DataBuffer::add_data (double timestamp, float *value)
     lock.unlock ();
 }
 
-void DataBuffer::get_chunk (size_t start, size_t size, double *ts_buf, float *data_buf)
+void DataBuffer::get_chunk (size_t start, size_t size, double *ts_buf, double *data_buf)
 {
     if (start + size < buffer_size)
     {
         memcpy (ts_buf, timestamps + start, size * sizeof (double));
-        memcpy (data_buf, data + start * num_samples, size * sizeof (float) * num_samples);
+        memcpy (data_buf, data + start * num_samples, size * sizeof (double) * num_samples);
     }
     else
     {
         size_t first_half = buffer_size - start;
         size_t second_half = size - first_half;
         memcpy (ts_buf, timestamps + start, first_half * sizeof (double));
-        memcpy (data_buf, data + start * num_samples, first_half * sizeof (float) * num_samples);
+        memcpy (data_buf, data + start * num_samples, first_half * sizeof (double) * num_samples);
         memcpy (ts_buf + first_half, timestamps, second_half * sizeof (double));
         memcpy (
-            data_buf + first_half * num_samples, data, second_half * sizeof (float) * num_samples);
+            data_buf + first_half * num_samples, data, second_half * sizeof (double) * num_samples);
     }
 }
 
 // removes data from buffer
-size_t DataBuffer::get_data (size_t max_count, double *ts_buf, float *data_buf)
+size_t DataBuffer::get_data (size_t max_count, double *ts_buf, double *data_buf)
 {
     lock.lock ();
     size_t result_count = max_count;
@@ -72,7 +72,7 @@ size_t DataBuffer::get_data (size_t max_count, double *ts_buf, float *data_buf)
 }
 
 // doesn't remove data from buffer
-size_t DataBuffer::get_current_data (size_t max_count, double *ts_buf, float *data_buf)
+size_t DataBuffer::get_current_data (size_t max_count, double *ts_buf, double *data_buf)
 {
     lock.lock ();
     size_t result_count = max_count;
