+++ src/utils/data_buffer.h
@@ -10,7 +10,7 @@ class DataBuffer
     SpinLock lock;
 
     double *timestamps;
-    float *data;
+    double *data;
 
     size_t buffer_size;
     size_t first_used, first_free;
@@ -22,15 +22,15 @@ class DataBuffer
         return (index + 1) % buffer_size;
     }
 
-    void get_chunk (size_t start, size_t size, double *tsBuf, float *data_buf);
+    void get_chunk (size_t start, size_t size, double *tsBuf, double *data_buf);
 
 public:
     DataBuffer (int num_samples, size_t buffer_size);
     ~DataBuffer ();
 
-    void add_data (double timestamp, float *value);
-    size_t get_data (size_t max_count, double *ts_buf, float *data_buf);
-    size_t get_current_data (size_t max_count, double *ts_buf, float *data_buf);
+    void add_data (double timestamp, double *value);
+    size_t get_data (size_t max_count, double *ts_buf, double *data_buf);
+    size_t get_current_data (size_t max_count, double *ts_buf, double *data_buf);
     size_t get_data_count ();
     bool is_ready ();
 };
